Splitting unit tests.
